requirements.txt
# AlphaAgentEvo dependencies
# Install: pip install -r requirements.txt

# PyTorch (install separately for CUDA support)
# pip install torch --index-url https://download.pytorch.org/whl/cu124
torch

# Transformers + RL training
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.12.0
datasets>=3.0.0

# API server
fastapi>=0.115.0
uvicorn>=0.30.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
h5py>=3.9.0
pyarrow>=14.0.0
pyparsing>=3.1.0
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
pyyaml>=6.0
scipy>=1.11.0

# Memory efficient training
bitsandbytes>=0.43.0

# Expression parsing dependency
joblib>=1.3.0
